# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Safe Place API",
    "version": "1.0.0"
  },
  "schemes": [
    "http"
  ],
  "host": "localhost:3000",
  "basePath": "/api",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/messages": {
      "get": {
        "summary": "Get all messages",
        "operationId": "getAllMessages",
        "tags": ["Messages"],
        "responses": {
          "200": {
            "description": "A list of messages",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Message"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new message",
        "operationId": "createMessage",
        "tags": ["Messages"],
        "parameters": [
          {
            "in": "body",
            "name": "message",
            "description": "Message object to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewMessage"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Message created successfully",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      }
    },
    "/messages/{id}": {
      "get": {
        "summary": "Get message by ID",
        "operationId": "getMessageById",
        "tags": ["Messages"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Message details",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "404": {
            "description": "Message not found"
          }
        }
      },
      "put": {
        "summary": "Update message by ID",
        "operationId": "updateMessage",
        "tags": ["Messages"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "message",
            "description": "Updated message object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewMessage"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message updated successfully",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "404": {
            "description": "Message not found"
          }
        }
      },
      "delete": {
        "summary": "Delete message by ID",
        "operationId": "deleteMessage",
        "tags": ["Messages"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Message deleted successfully"
          },
          "404": {
            "description": "Message not found"
          }
        }
      }
    },
...
    "/friends": {
      "get": {
        "summary": "Get all friendships",
        "operationId": "getAllFriends",
        "tags": ["Friends"],
        "responses": {
          "200": {
            "description": "A list of friendships",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Friend"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new friendship",
        "operationId": "createFriend",
        "tags": ["Friends"],
        "parameters": [
          {
            "in": "body",
            "name": "friend",
            "description": "Friendship object to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewFriend"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Friendship created successfully",
            "schema": {
              "$ref": "#/definitions/Friend"
            }
          }
        }
      }
    },
    "/friends/{id_user}/{id_friend}": {
      "get": {
        "summary": "Get friendship by user and friend IDs",
        "operationId": "getFriendById",
        "tags": ["Friends"],
        "parameters": [
          {
            "name": "id_user",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id_friend",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Friendship details",
            "schema": {
              "$ref": "#/definitions/Friend"
            }
          },
          "404": {
            "description": "Friendship not found"
          }
        }
      },
      "put": {
        "summary": "Update friendship by user and friend IDs",
        "operationId": "updateFriend",
        "tags": ["Friends"],
        "parameters": [
          {
            "name": "id_user",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id_friend",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "friend",
            "description": "Updated friendship object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewFriend"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Friendship updated successfully",
            "schema": {
              "$ref": "#/definitions/Friend"
            }
          },
          "404": {
            "description": "Friendship not found"
          }
        }
      },
      "delete": {
        "summary": "Delete friendship by user and friend IDs",
        "operationId": "deleteFriend",
        "tags": ["Friends"],
        "parameters": [
          {
            "name": "id_user",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id_friend",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Friendship deleted successfully"
          },
          "404": {
            "description": "Friendship not found"
          }
        }
      }
    },
...
    "/discussions": {
      "get": {
        "summary": "Get all discussions (associations between users and conversations)",
        "operationId": "getAllDiscussions",
        "tags": ["Discussions"],
        "responses": {
          "200": {
            "description": "A list of discussions",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Discussion"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new discussion (association between a user and a conversation)",
        "operationId": "createDiscussion",
        "tags": ["Discussions"],
        "parameters": [
          {
            "in": "body",
            "name": "discussion",
            "description": "Discussion object to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewDiscussion"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Discussion created successfully",
            "schema": {
              "$ref": "#/definitions/Discussion"
            }
          }
        }
      }
    },
    "/discussions/{id_user}/{id_conversation}": {
      "get": {
        "summary": "Get discussion by User and Conversation IDs",
        "operationId": "getDiscussionById",
        "tags": ["Discussions"],
        "parameters": [
          {
            "name": "id_user",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id_conversation",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Discussion details",
            "schema": {
              "$ref": "#/definitions/Discussion"
            }
          },
          "404": {
            "description": "Discussion not found"
          }
        }
      },
      "put": {
        "summary": "Update discussion by User and Conversation IDs",
        "operationId": "updateDiscussion",
        "tags": ["Discussions"],
        "parameters": [
          {
            "name": "id_user",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id_conversation",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "discussion",
            "description": "Updated discussion object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewDiscussion"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discussion updated successfully",
            "schema": {
              "$ref": "#/definitions/Discussion"
            }
          },
          "404": {
            "description": "Discussion not found"
          }
        }
      },
      "delete": {
        "summary": "Delete discussion by User and Conversation IDs",
        "operationId": "deleteDiscussion",
        "tags": ["Discussions"],
        "parameters": [
          {
            "name": "id_user",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id_conversation",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Discussion deleted successfully"
          },
          "404": {
            "description": "Discussion not found"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "Id_User": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "first_name": {
          "type": "string"
        },
        "birth_date": {
          "type": "string",
          "format": "date-time"
        },
        "mail": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "isAdmin": {
          "type": "boolean"
        },
        "seed_totp": {
          "type": "string"
        }
      },
      "required": ["last_name", "first_name", "birth_date", "mail", "password"]
    },
...
    "Message": {
      "type": "object",
      "properties": {
        "Id_Message": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "date_upload": {
          "type": "string",
          "format": "date-time"
        },
        "Id_User": {
          "type": "string"
        },
        "Id_Conversation": {
          "type": "string",
          "description": "ID of the conversation this message belongs to"
        }
      },
      "required": ["content", "date_upload", "Id_User"]
    },
    "NewMessage": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "date_upload": {
          "type": "string",
          "format": "date-time"
        },
        "Id_User": {
          "type": "string"
        },
        "Id_Conversation": {
          "type": "string",
          "description": "ID of the conversation this message belongs to"
        }
      },
      "required": ["content", "date_upload", "Id_User"]
    },
...
    "Conversation": {
      "type": "object",
      "properties": {
        "Id_Conversation": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "Discuss": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Discuss"
          }
        },
        "Message": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Message"
          }
        }
      },
      "required": ["Id_Conversation"]
    },
...
    "Friend": {
      "type": "object",
      "properties": {
        "Id_User": {
          "type": "string"
        },
        "Id_Friend": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "friend": {
          "$ref": "#/definitions/User"
        }
      },
      "required": ["Id_User", "Id_Friend"]
    },
    "NewFriend": {
      "type": "object",
      "properties": {
        "Id_User": {
          "type": "string"
        },
        "Id_Friend": {
          "type": "string"
        }
      }
    },
...
    "Discussion": {
      "type": "object",
      "properties": {
        "Id_User": {
          "type": "string"
        },
        "Id_Conversation": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "conversation": {
          "$ref": "#/definitions/Conversation"
        }
      },
      "required": ["Id_User", "Id_Conversation"]
    },
    "NewDiscussion": {
      "type": "object",
      "properties": {
        "Id_User": {
          "type": "string"
        },
        "Id_Conversation": {
          "type": "string"
        }
      }
    }
  }
}